Added second TxBody spec example. All Conway HasSpec instances and EraSpecXX instances are exported   by Test.Cardano.Ledger.Constrained.Conway.Instances Added LedgerType.Tests
